fix: adjust page naming
