+++ app/Filament/Resources/BastProjectResource/Pages/ListHomeConnectDetails.php
@@ -4,7 +4,7 @@ namespace App\Filament\Resources\BastProjectResource\Pages;
 
 use App\Filament\Resources\BastProjectResource;
 use App\Models\BastProject;
-use App\Models\ODPDetail;
+use App\Models\HomeConnect;
 use Filament\Pages\Page;
 use Filament\Tables;
 use Filament\Tables\Columns\TextColumn;
@@ -15,72 +15,77 @@ use Filament\Resources\Pages\ListRecords;
 use Illuminate\Database\Eloquent\Builder;
 use Maatwebsite\Excel\Facades\Excel;
 use Filament\Tables\Actions\Action;
-use App\Exports\BastODPExport;
+use App\Exports\BastPoleExport;
 use Filament\Tables\Columns\ImageColumn;
+use Filament\Tables\Filters\SelectFilter;
+use App\Services\DownloadBAService;
 use Illuminate\Support\Facades\Auth;
 use Filament\Forms\Form as FilamentForm;
 
-class listOdpDetails extends ListRecords
+
+class ListHomeConnectDetails extends ListRecords
 {
     use InteractsWithTable;
     protected static string $resource = BastProjectResource::class;
 
-    protected static ?string $title = 'List ODP Details';
+    protected static ?string $title = 'List Home Connect Details';
 
-    protected static ?string $navigationLabel = 'ODP Details';
+    protected static ?string $navigationLabel = 'Home Connect Details';
     protected static ?string $navigationIcon = 'heroicon-o-collection';
-    protected static ?string $slug = 'list-odp-details';
-    public ?string $site = null;
+    protected static ?string $slug = 'list-homeconnect-details';
+     public ?string  $site = null;
 
-    public function mount(?string $site = null): void
+     public function mount(?string $site = null): void
     {
         $this->site = $site;
     }
 
     protected function getTableQuery(): Builder
     {
-        return ODPDetail::query()
-                ->Join('BastProject', 'ODPDetail.bast_id', '=', 'BastProject.bast_id')
-                ->when($this->site, fn($query) => 
-                        $query->where('BastProject.site', $this->site)
-                    )
-                    ->select('ODPDetail.*', 'BastProject.site');
+        return HomeConnect::query()
+                ->Join('BastProject', 'HomeConnect.bast_id', '=', 'BastProject.bast_id')
+                ->where('HomeConnect.site', $this->site)
+                    ->select('HomeConnect.*', 'BastProject.site');
     }
 
     protected function getTableColumns(): array
     {
         return [
-            TextColumn::make('bast_id')->label('BAST ID')->searchable(['BastProject.bast_id']),
+            // TextColumn::make('bast_id')->label('BAST ID')->searchable(),
             TextColumn::make('site')->label('Site')->searchable(['BastProject.site']),
-            TextColumn::make('odc_name')->searchable(['ODPDetail.odc_name']),
-            TextColumn::make('odp_name')->searchable(['ODPDetail.odp_name']),
-            TextColumn::make('notes')->searchable(['ODPDetail.notes']),
-            ImageColumn::make('instalasi')
-                ->label('Instalasi')
+            TextColumn::make('odp_name')->label('ODP')->searchable(['HomeConnect.odp_name']),
+            TextColumn::make('port_odp')->label('Port ODP')->searchable(['HomeConnect.port_odp']),
+            TextColumn::make('status_port')->label('Status')->searchable(['HomeConnect.status_port']),
+            TextColumn::make('merk_ont')->label('Merk ONT')->searchable(['HomeConnect.merk_ont']),
+            TextColumn::make('sn_ont')->label('SN ONT')->searchable(['HomeConnect.sn_ont']),
+            TextColumn::make('notes')->searchable(['HomeConnect.notes']),
+
+            ImageColumn::make('foto_label_id_plg')
+                ->label('ID Pelanggan di ODP')
                 ->disk('public')
-                ->getStateUsing(fn($record) => $record->instalasi ? asset('storage/'.$record->instalasi) : null)
-                ->width(150)
-                ->height(150),
-            ImageColumn::make('odp_terbuka')
-                ->label('ODP Terbuka')
-                ->disk('public')
-                ->getStateUsing(fn($record) => $record->odp_terbuka ? asset('storage/'.$record->odp_terbuka) : null)
-                ->width(150)
-                ->height(150),
-            ImageColumn::make('odp_tertutup')
-                ->label('ODP Tertutup')
-                ->disk('public')
-                ->getStateUsing(fn($record) => $record->odp_tertutup ? asset('storage/'.$record->odp_tertutup) : null)
+                ->getStateUsing(fn($record) => $record->foto_label_id_plg ? asset('storage/'.$record->foto_label_id_plg) : null)
                 ->width(150)
                 ->height(150),
 
-            ImageColumn::make('power_optic_odc')
-                ->label('Power Optic ODC')
+            ImageColumn::make('foto_qr')
+                ->label('QR Code')
                 ->disk('public')
-                ->getStateUsing(fn($record) => $record->power_optic_odc ? asset('storage/'.$record->power_optic_odc) : null)
+                ->getStateUsing(fn($record) => $record->foto_qr ? asset('storage/'.$record->foto_qr) : null)
+                ->width(150)
+                ->height(150),
+            ImageColumn::make('foto_label_odp')
+                ->label('ODP')
+                ->disk('public')
+                ->getStateUsing(fn($record) => $record->foto_label_odp ? asset('storage/'.$record->foto_label_odp) : null)
                 ->width(150)
                 ->height(150),
 
+            ImageColumn::make('foto_sn_ont')
+                ->label('SN ONT')
+                ->disk('public')
+                ->getStateUsing(fn($record) => $record->foto_sn_ont ? asset('storage/'.$record->foto_sn_ont) : null)
+                ->width(150)
+                ->height(150),
             TextColumn::make('progress_percentage')
                 ->label('Progress (%)')
                 ->formatStateUsing(fn ($state) => '
@@ -91,7 +96,7 @@ class listOdpDetails extends ListRecords
                     </div>
                     <div style="font-size:12px; text-align:center; margin-top:2px;">'.number_format($state,0).'%</div>
                 ')
-                ->html() 
+                ->html()
                 ->sortable(),
             TextColumn::make('status')
                     ->badge()
@@ -111,6 +116,7 @@ class listOdpDetails extends ListRecords
                     })
                     ->sortable(),
         ];
+
     }
 
     protected function getTableActions(): array
@@ -129,7 +135,7 @@ class listOdpDetails extends ListRecords
             //     ->requiresConfirmation()
             //     ->visible(fn ($record) => $record->status === 'submit'  && (int) $record->progress_percentage >= 100)
             //     ->action(function ($record) {
-            //         ODPDetail::where('bast_id', $record->bast_id)->where('odp_name', $record->odp_name)
+            //         HomeConnect::where('bast_id', $record->bast_id)->where('id_pelanggan', $record->id_pelanggan)
             //         ->update([
             //             'status'       => 'pending',
             //             'approval_by'  => Auth::user()->email,
@@ -144,7 +150,7 @@ class listOdpDetails extends ListRecords
                 ->requiresConfirmation()
                 ->visible(fn ($record) => $record->status === 'submit' && (int) $record->progress_percentage >= 100)
                 ->action(function ($record) {
-                    ODPDetail::where('bast_id', $record->bast_id)->where('odp_name', $record->odp_name)
+                    HomeConnect::where('bast_id', $record->bast_id)->where('id_pelanggan', $record->id_pelanggan)
                     ->update([
                         'status'       => 'approved',
                         'approval_by'  => Auth::user()->email,
@@ -152,7 +158,7 @@ class listOdpDetails extends ListRecords
                     ]);
                 })->after(fn () => $this->dispatch('refresh'))
                 ->successNotificationTitle('Data berhasil di-approve'),
-            Action::make('reject')
+             Action::make('reject')
                 ->label('Rejected')
                 ->icon('heroicon-o-x-circle')
                 ->color('danger')
@@ -167,28 +173,28 @@ class listOdpDetails extends ListRecords
                 ])
 
                 ->mountUsing(function (FilamentForm  $form, $record) {
-                    $detail = ODPDetail::where('bast_id', $record->bast_id)
-                        ->where('odp_name', $record->odp_name)
+                    $detail = HomeConnect::where('bast_id', $record->bast_id)
+                        ->where('id_pelanggan', $record->id_pelanggan)
                         ->first();
 
                     $form->fill([
                         'notes' => '',
-                        'old_notes' => $detail?->notes, 
+                        'old_notes' => $detail?->notes,
                     ]);
                 })
 
                 ->action(function ($record, array $data) {
-                    $detail = ODPDetail::where('bast_id', $record->bast_id)
-                        ->where('odp_name', $record->odp_name)
+                    $detail = HomeConnect::where('bast_id', $record->bast_id)
+                        ->where('id_pelanggan', $record->id_pelanggan)
                         ->first();
 
                     $oldNotes = $detail?->notes ?? '';
                     $newEntry = "[" . now() . "] " . Auth::user()->email . "-> Reject :\n" .
                                 $data['notes'] . "\n\n";
-                    $finalNotes = $newEntry  ." | ". $oldNotes;
+                    $finalNotes = $newEntry ." | " . $oldNotes;
 
-                    ODPDetail::where('bast_id', $record->bast_id)
-                        ->where('odp_name', $record->odp_name)
+                    HomeConnect::where('bast_id', $record->bast_id)
+                        ->where('id_pelanggan', $record->id_pelanggan)
                         ->update([
                             'status'       => 'rejected',
                             'approval_by'  => Auth::user()->email,
@@ -199,14 +205,79 @@ class listOdpDetails extends ListRecords
 
                 ->after(fn () => $this->dispatch('refresh'))
                 ->successNotificationTitle('Data berhasil di-reject'),
-            // Tables\Actions\ViewAction::make(),
-            Action::make('export_implementation')
-                    ->label('Print')
-                    ->icon('heroicon-o-document-arrow-down')
-                    ->action(fn ($record) => Excel::download(new BastODPExport($record), "Implementation_ODP_{$record->odp_name}.xlsx"))
-                    ->visible(fn ($record) => $record->status === 'approved'),
+            Action::make('export_ba_pdf')
+                ->label('Export BA PDF')
+                ->icon('heroicon-o-arrow-down-tray')
+                ->color('danger')
+                ->action(function ($record) {
+                    $pdf = app(DownloadBAService::class)->downloadBA($record);
+
+                    // return response()->streamDownload(
+                    //     fn () => print($pdf->output()),
+                    //     "BA_{$record->site}.pdf"
+                    // );
+                    return response()->streamDownload(
+                    fn () =>print($pdf->output()),
+                    "BA_Pelanggan_{$record->id_pelanggan}.pdf",
+                    [
+                        "Content-Type" => "application/pdf",
+                    ]
+                );
+                })->visible(fn ($record) => $record->status === 'approved'),
         ];
     }
+
+    protected function getTableFilters(): array
+    {
+        return [
+            SelectFilter::make('odp_name')
+                ->label('ODP Name')
+                ->options(
+                    HomeConnect::query()
+                        ->when($this->site, fn($q) => $q->where('site', $this->site))
+                        ->whereNotNull('odp_name')
+                        ->distinct()
+                        ->orderBy('odp_name')
+                        ->pluck('odp_name', 'odp_name')
+                        ->filter() // buang null/empty
+                        ->mapWithKeys(fn($v) => [(string)$v => (string)$v]) // pastikan string keys
+                        ->toArray()
+                )
+                ->searchable(),
+            SelectFilter::make('status_port')
+                ->label('Status Port')
+                ->options(
+                    HomeConnect::query()
+                        ->when($this->site, fn($q) => $q->where('site', $this->site))
+                        ->whereNotNull('status_port')
+                        ->distinct()
+                        ->orderBy('status_port')
+                        ->pluck('status_port', 'status_port')
+                        ->filter()
+                        ->mapWithKeys(fn($v) => [(string)$v => (string)$v])
+                        ->toArray()
+                )
+                ->searchable(),
+
+            SelectFilter::make('merk_ont')
+                ->label('Merk ONT')
+                ->options(
+                    HomeConnect::query()
+                        ->when($this->site, fn($q) => $q->where('site', $this->site))
+                        ->whereNotNull('merk_ont')
+                        ->distinct()
+                        ->orderBy('merk_ont')
+                        ->pluck('merk_ont', 'merk_ont')
+                        ->filter()
+                        ->mapWithKeys(fn($v) => [(string)$v => (string)$v])
+                        ->toArray()
+                )
+                ->searchable(),
+
+        ];
+    }
+
+
 
     protected function getTableBulkActions(): array
     {
